facial_recognition_system/requirements.txt
# requirements.txt
# Sistema de Reconhecimento Facial - Dependências

# Visão computacional e reconhecimento facial
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2

# Processamento numérico
numpy==1.24.3

# Banco de dados MySQL
mysql-connector-python==8.1.0

# Processamento de imagens
Pillow==10.0.0

# Utilitários opcionais
python-dotenv==1.0.0

# Para instalação no Windows (se necessário)
# cmake==3.27.0
# wheel==0.41.0